+++ inlang/source-code/plugins/json/src/plugin.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import type { Message, Variant, LanguageTag, Plugin, NodeishFilesystemSubset } from "@inlang/sdk"
-import { id, displayName, description } from "../marketplace-manifest.json"
+import { displayName, description } from "../marketplace-manifest.json"
 import { throwIfInvalidSettings, type PluginSettings } from "./settings.js"
 import { replaceAll } from "./utilities.js"
 import { ideExtensionConfig } from "./ideExtension/config.js"
@@ -19,26 +19,30 @@ import {
  */
 let DETECTFORMATTING: DetectJsonFormattingApi = detectJsonFormatting()
 
-export const plugin: Plugin<PluginSettings> = {
-	id: id as Plugin["id"],
+const id = "plugin.inlang.json"
+
+export const plugin: Plugin<{
+	[id]: PluginSettings
+}> = {
+	id,
 	displayName,
 	description,
-	loadMessages: async ({ languageTags, sourceLanguageTag, settings, nodeishFs }) => {
-		settings.variableReferencePattern = settings.variableReferencePattern || ["{", "}"]
-		throwIfInvalidSettings(settings)
+	loadMessages: async ({ settings, nodeishFs }) => {
+		settings[id].variableReferencePattern = settings[id].variableReferencePattern || ["{", "}"]
+		throwIfInvalidSettings(settings[id])
 		return loadMessages({
 			nodeishFs,
-			settings,
-			languageTags,
-			sourceLanguageTag,
+			pluginSettings: settings[id],
+			languageTags: settings.languageTags,
+			sourceLanguageTag: settings.sourceLanguageTag,
 		})
 	},
 	saveMessages: async ({ messages, settings, nodeishFs }) => {
-		settings.variableReferencePattern = settings.variableReferencePattern || ["{", "}"]
-		throwIfInvalidSettings(settings)
+		settings[id].variableReferencePattern = settings[id].variableReferencePattern || ["{", "}"]
+		throwIfInvalidSettings(settings[id])
 		return saveMessages({
 			nodeishFs,
-			settings,
+			pluginSettings: settings[id],
 			messages,
 		})
 	},
@@ -52,31 +56,31 @@ export const plugin: Plugin<PluginSettings> = {
  */
 async function loadMessages(args: {
 	nodeishFs: NodeishFilesystemSubset
-	settings: PluginSettings
+	pluginSettings: PluginSettings
 	languageTags: Readonly<LanguageTag[]>
 	sourceLanguageTag: string
 }): Promise<Message[]> {
 	const messages: Message[] = []
 	for (const languageTag of resolveOrderOfLanguageTags(args.languageTags, args.sourceLanguageTag)) {
-		if (typeof args.settings.pathPattern !== "string") {
-			for (const [prefix, path] of Object.entries(args.settings.pathPattern)) {
+		if (typeof args.pluginSettings.pathPattern !== "string") {
+			for (const [prefix, path] of Object.entries(args.pluginSettings.pathPattern)) {
 				const messagesFromFile = await getFileToParse(
 					path,
 					languageTag,
 					args.sourceLanguageTag,
-					args.nodeishFs,
+					args.nodeishFs
 				)
 				for (const [key, value] of Object.entries(messagesFromFile)) {
 					const prefixedKey = prefix + ":" + replaceAll(key, "u002E", ".")
-					addVariantToMessages(messages, prefixedKey, languageTag, value, args.settings)
+					addVariantToMessages(messages, prefixedKey, languageTag, value, args.pluginSettings)
 				}
 			}
 		} else {
 			const messagesFromFile = await getFileToParse(
-				args.settings.pathPattern,
+				args.pluginSettings.pathPattern,
 				languageTag,
 				args.sourceLanguageTag,
-				args.nodeishFs,
+				args.nodeishFs
 			)
 			for (const [key, value] of Object.entries(messagesFromFile)) {
 				addVariantToMessages(
@@ -84,7 +88,7 @@ async function loadMessages(args: {
 					replaceAll(key, "u002E", "."),
 					languageTag,
 					value,
-					args.settings,
+					args.pluginSettings
 				)
 			}
 		}
@@ -103,7 +107,7 @@ async function getFileToParse(
 	path: string,
 	languageTag: string,
 	sourceLanguageTag: string,
-	nodeishFs: NodeishFilesystemSubset,
+	nodeishFs: NodeishFilesystemSubset
 ): Promise<Record<string, string>> {
 	const pathWithLanguage = path.replace("{languageTag}", languageTag)
 	// get file, make sure that is not braking when the namespace doesn't exist in every languageTag dir
@@ -138,7 +142,7 @@ async function getFileToParse(
  */
 const resolveOrderOfLanguageTags = (
 	languageTags: Readonly<LanguageTag[]>,
-	sourceLanguageTag: LanguageTag,
+	sourceLanguageTag: LanguageTag
 ): LanguageTag[] => {
 	const filteredTags = languageTags.filter((t) => t !== sourceLanguageTag) // Remove sourceLanguageTag
 	filteredTags.unshift(sourceLanguageTag) // Add sourceLanguageTag to the beginning of the filtered array
@@ -155,13 +159,13 @@ const addVariantToMessages = (
 	key: string,
 	languageTag: LanguageTag,
 	value: string,
-	settings: PluginSettings,
+	settings: PluginSettings
 ) => {
 	const messageIndex = messages.findIndex((m) => m.id === key)
 	if (messageIndex !== -1) {
 		const variant: Variant = {
 			languageTag,
-			match: {},
+			match: [],
 			pattern: parsePattern(value, settings.variableReferencePattern!),
 		}
 
@@ -177,7 +181,7 @@ const addVariantToMessages = (
 		message.variants = [
 			{
 				languageTag,
-				match: {},
+				match: [],
 				pattern: parsePattern(value, settings.variableReferencePattern!),
 			},
 		]
@@ -196,7 +200,7 @@ function parsePattern(text: string, variableReferencePattern: string[]): Variant
 	const expression = variableReferencePattern[1]
 		? new RegExp(
 				`(\\${variableReferencePattern[0]}[^\\${variableReferencePattern[1]}]+\\${variableReferencePattern[1]})`,
-				"g",
+				"g"
 		  )
 		: new RegExp(`(${variableReferencePattern}\\w+)`, "g")
 	const pattern: Variant["pattern"] = text
@@ -210,7 +214,7 @@ function parsePattern(text: string, variableReferencePattern: string[]): Variant
 						? element.slice(
 								variableReferencePattern[0].length,
 								// negative index, removing the trailing pattern
-								-variableReferencePattern[1].length,
+								-variableReferencePattern[1].length
 						  )
 						: element.slice(variableReferencePattern[0].length),
 				}
@@ -232,10 +236,10 @@ function parsePattern(text: string, variableReferencePattern: string[]): Variant
  */
 async function saveMessages(args: {
 	nodeishFs: NodeishFilesystemSubset
-	settings: PluginSettings
+	pluginSettings: PluginSettings
 	messages: Message[]
 }) {
-	if (typeof args.settings.pathPattern === "object") {
+	if (typeof args.pluginSettings.pathPattern === "object") {
 		// with namespaces
 		const storage: Record<
 			LanguageTag,
@@ -245,7 +249,7 @@ async function saveMessages(args: {
 			for (const variant of message.variants) {
 				const prefix: string = message.id.includes(":")
 					? (message.id.split(":")[0] as string)
-					: (Object.keys(args.settings.pathPattern)[0] as string)
+					: (Object.keys(args.pluginSettings.pathPattern)[0] as string)
 				const resolvedId = message.id.replace(prefix + ":", "")
 
 				storage[variant.languageTag] ??= {}
@@ -254,7 +258,7 @@ async function saveMessages(args: {
 			}
 		}
 		for (const [languageTag, _value] of Object.entries(storage)) {
-			for (const path of Object.values(args.settings.pathPattern)) {
+			for (const path of Object.values(args.pluginSettings.pathPattern)) {
 				// check if directory exists
 				const directoryPath = path
 					.replace("{languageTag}", languageTag)
@@ -268,13 +272,13 @@ async function saveMessages(args: {
 				}
 			}
 			for (const [prefix, value] of Object.entries(_value)) {
-				const pathWithLanguage = (args.settings.pathPattern[prefix] as string).replace(
+				const pathWithLanguage = (args.pluginSettings.pathPattern[prefix] as string).replace(
 					"{languageTag}",
-					languageTag,
+					languageTag
 				)
 				await args.nodeishFs.writeFile(
 					pathWithLanguage,
-					serializeFile(value, args.settings.variableReferencePattern),
+					serializeFile(value, args.pluginSettings.variableReferencePattern)
 				)
 			}
 		}
@@ -288,7 +292,7 @@ async function saveMessages(args: {
 			}
 		}
 		for (const [languageTag, value] of Object.entries(storage)) {
-			const pathWithLanguage = args.settings.pathPattern.replace("{languageTag}", languageTag)
+			const pathWithLanguage = args.pluginSettings.pathPattern.replace("{languageTag}", languageTag)
 			try {
 				await args.nodeishFs.readdir(pathWithLanguage.split("/").slice(0, -1).join("/"))
 			} catch {
@@ -299,7 +303,7 @@ async function saveMessages(args: {
 
 			await args.nodeishFs.writeFile(
 				pathWithLanguage,
-				serializeFile(value, args.settings.variableReferencePattern),
+				serializeFile(value, args.pluginSettings.variableReferencePattern)
 			)
 		}
 	}
@@ -314,7 +318,7 @@ async function saveMessages(args: {
  */
 function serializeFile(
 	messages: Record<Message["id"], Variant["pattern"]>,
-	variableReferencePattern: PluginSettings["variableReferencePattern"],
+	variableReferencePattern: PluginSettings["variableReferencePattern"]
 ): string {
 	let result: Record<string, string> = {}
 	for (const [messageId, pattern] of Object.entries(messages)) {
@@ -352,7 +356,7 @@ function serializePattern(pattern: Variant["pattern"], variableReferencePattern:
 				result.push(
 					variableReferencePattern[1]
 						? `${variableReferencePattern[0]}${element.name}${variableReferencePattern[1]}`
-						: `${variableReferencePattern[0]}${element.name}`,
+						: `${variableReferencePattern[0]}${element.name}`
 				)
 				break
 			default: