+++ inlang/source-code/plugins/i18next/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inlang/plugin-i18next
 
+## 4.6.0
+
+### Minor Changes
+
+- 6e4ea967d: refactor: now uses the plugin api v2.0
+
+### Patch Changes
+
+- Updated dependencies [0f9dc72b3]
+  - @inlang/sdk@0.9.0
+
+## 4.5.0
+
+### Minor Changes
+
+- b7dfc781e: change message format match from object to array
+
+### Patch Changes
+
+- Updated dependencies [b7dfc781e]
+  - @inlang/sdk@0.8.0
+
 ## 4.4.0
 
 ### Minor Changes